[1.12] Updates to 1.4.4
Wagon, many bug fixes

+++ src/main/java/com/animania/common/entities/goats/ai/EntityAIFindFoodGoats.java
@@ -9,11 +9,14 @@ import com.animania.config.AnimaniaConfig;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockCrops;
+import net.minecraft.block.BlockFlower;
 import net.minecraft.entity.EntityCreature;
 import net.minecraft.entity.ai.EntityAIBase;
+import net.minecraft.entity.ai.RandomPositionGenerator;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3d;
 
 public class EntityAIFindFoodGoats extends EntityAIBase 
 {
@@ -49,6 +52,17 @@ public class EntityAIFindFoodGoats extends EntityAIBase
 					return false;		
 				}
 			} 
+			
+			if (this.temptedEntity.getRNG().nextInt(100) == 0)
+			{
+				Vec3d vec3d = RandomPositionGenerator.findRandomTarget(this.temptedEntity, 20, 4);
+				if (vec3d != null) {
+					this.delayTemptCounter = 0;
+					this.resetTask();
+					this.temptedEntity.getNavigator().tryMoveToXYZ(vec3d.x, vec3d.y, vec3d.z, this.speed);
+				}
+				return false;
+			}
 			
 			BlockPos currentpos = new BlockPos(temptedEntity.posX, temptedEntity.posY, temptedEntity.posZ);
 			BlockPos trypos1 = new BlockPos(temptedEntity.posX + 1, temptedEntity.posY, temptedEntity.posZ);
@@ -105,7 +119,7 @@ public class EntityAIFindFoodGoats extends EntityAIBase
 				} 
 			}
 
-			if (poschk == Blocks.RED_FLOWER || poschk instanceof BlockCrops|| poschk == Blocks.WHEAT || poschk == Blocks.YELLOW_FLOWER) {
+			if (poschk == Blocks.RED_FLOWER || poschk instanceof BlockCrops|| poschk == Blocks.WHEAT || poschk instanceof BlockFlower) {
 
 				if (temptedEntity instanceof EntityAnimaniaGoat) {
 					EntityAnimaniaGoat ech = (EntityAnimaniaGoat)temptedEntity;
@@ -114,7 +128,10 @@ public class EntityAIFindFoodGoats extends EntityAIBase
 				} 
 				
 				if (AnimaniaConfig.gameRules.plantsRemovedAfterEating) {
-					temptedEntity.world.destroyBlock(currentpos, false);
+					Block destchk = temptedEntity.world.getBlockState(currentpos).getBlock();
+					if (destchk != BlockHandler.blockTrough) {
+						temptedEntity.world.destroyBlock(currentpos, false);
+					}
 				}
 				this.delayTemptCounter = 0;
 				return false;
